txnfeed: use source side ordering
This deletes the destination side ordered feed and switches to using the
ordered feed added by #164558.

Validation was added to the merge feed to ensure data from the source
cluster is exported in the expected order.

Release note: none
Informs: CRDB-60143
