+++ pkg/crosscluster/logical/txnfeed/BUILD.bazel
@@ -5,7 +5,6 @@ go_library(
     srcs = [
         "merge_feed.go",
         "merge_heap.go",
-        "ordered_feed.go",
         "transactions.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/crosscluster/logical/txnfeed",
@@ -18,7 +17,6 @@ go_library(
         "//pkg/roachpb",
         "//pkg/util/ctxgroup",
         "//pkg/util/hlc",
-        "//pkg/util/span",
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_cockroachdb_redact//:redact",
     ],
@@ -29,7 +27,7 @@ go_test(
     srcs = [
         "merge_feed_bench_test.go",
         "merge_feed_test.go",
-        "ordered_feed_test.go",
+        "test_helpers_test.go",
         "transactions_test.go",
     ],
     embed = [":txnfeed"],
@@ -41,7 +39,6 @@ go_test(
         "//pkg/roachpb",
         "//pkg/util/ctxgroup",
         "//pkg/util/hlc",
-        "//pkg/util/span",
         "//pkg/util/timeutil",
         "@com_github_stretchr_testify//require",
     ],
